requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
passlib
bcrypt
pyjwt
pydantic
email-validator
python-dotenv
uuid
shortuuid